Added virtual memory in assembly Doesn't work with secondary cores yet

+++ src/drivers/Gpio.cpp
@@ -11,7 +11,7 @@ void GPIO::unload(){};
 const char* GPIO::getName() { return "GPIO Driver\n"; }
 
 unsigned int GPIO::gpio_call(unsigned int pin_number, unsigned int value,
-                             unsigned int base, unsigned int field_size,
+                             unsigned long base, unsigned int field_size,
                              unsigned int field_max) {
   unsigned int field_mask = (1 << field_size) - 1;
 
@@ -19,7 +19,7 @@ unsigned int GPIO::gpio_call(unsigned int pin_number, unsigned int value,
   if (value > field_mask) return 0;
 
   unsigned int num_fields = 32 / field_size;
-  unsigned int reg = base + ((pin_number / num_fields) * 4);
+  unsigned long reg = base + ((pin_number / num_fields) * 4);
   unsigned int shift = (pin_number % num_fields) * field_size;
 
   unsigned int curval = MMIO::read(reg);
